Solution to Check Profit or Loss

+++ Check Profit or Loss@@6788dd63490637cfc44e1b98/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 int main() {
-    int a;
-    int b;
-    scanf("%d%d",&a,&b);
+    float a;
+    float b;
+    scanf("%f%f",&a,&b);
     if(a>b)
     printf("Profit");
     else if(a<b)
